Fix: 사용자 검색창 수정
